+++ examples/before.js
@@ -4,9 +4,9 @@
  *
  * Execute the function passed as parameter before the method, getter, setter or accessor.
  * The function passed as parameter receives all parameter passed to the target function.
- * If the function return false, the method, getter or setter isn't called
+ * If the function returns false, the method, getter or setter isn't called
  *
- * @param {function} fn
+ * @param {function({array}): {void|boolean}} fn - function called before the method
  * @returns {decorator}
  *
  */
@@ -38,7 +38,7 @@ function before (fn) {
     
     function _before(prev, next) {
       return function (...args) {
-        if (prev.call (this, ...args) !== false) {
+        if (prev.call (this, args) !== false) {
           return next.call (this, ...args);
         }
       };
@@ -46,32 +46,32 @@ function before (fn) {
   };
 }
 
-class C {
-  #p = 0;
-  
-  @init:before (() => console.log ('before get p'))
-  get p () {
-    return this.#p;
-  }
-  
-  @init:before ((v) => v > 10)
-  set p (v) {
-    return this.#p = v;
-  }
-  
-  @init:before (() => console.log ('before hello'))
-  hello () {
-    return 'hello world';
-  }
-  
-  @before(() => console.log('before accessor')) accessor
-  r = 1;
-  
-}
+// class C {
+//   #p = 0;
+//
+//   @init:before (() => console.log ('before get p'))
+//   get p () {
+//     return this.#p;
+//   }
+//
+//   @init:before ((v) => v > 10)
+//   set p (v) {
+//     return this.#p = v;
+//   }
+//
+//   @init:before (() => console.log ('before hello'))
+//   hello () {
+//     return 'hello world';
+//   }
+//
+//   @before(() => console.log('before accessor')) accessor
+//   r = 1;
+//
+// }
 
-const c = new C ();
-c.p     = 10;
-console.log (c.p);
-c.r = 20;
-console.log (c.r);
-console.log (c.hello ());
+// const c = new C ();
+// c.p     = 10;
+// console.log (c.p);
+// c.r = 20;
+// console.log (c.r);
+// console.log (c.hello ());