+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # JavascriptDecorators.org - CHANGELOG
 
+## version 0.1.2 - 06/02/2021
+
+- Fixed an error with accessor and static keywords. The previous version accept `accesor static`,
+  and it is `static accessor`.
+
 ## version 0.1.1 - 05/02/2021
 
 - Fixed an error with accessor transpilation when the decorator don't return an initialize function.